+++ people-and-teams/teams-client/src/services/legion-client/index.ts
@@ -695,6 +695,12 @@ export class LegionClient extends RestClient implements LegionClient {
 			largeAvatarImageUrl: teamResponse.largeAvatarImageUrl,
 			scopeMode: teamResponse.scopeMode,
 			isVerified: teamResponse.isVerified,
+			...(teamResponse.externalReference && {
+				externalReference: {
+					id: teamResponse.externalReference.id,
+					source: teamResponse.externalReference.source,
+				},
+			}),
 		};
 	}